refactor: move main to program.ts
