--- src/index.test.ts
@@ -1,8 +1,0 @@
-import { run } from './index'
-import { log } from '@eliasnorrby/log-util'
-
-test('it prints a message', () => {
-  log.info = jest.fn()
-  run()
-  expect(log.info).toHaveBeenCalledTimes(2)
-})
